Instantiate ov7670 device even if it is missing
The dts for trikboard declares two endpoints for vpif, and vpif
waits for both endpoints to be ready. Thus, if user plugged in
only one camera, vpif never gets initialized and special devices
/dev/video* are never created.

To mitigate this issue we hack around the endpoint driver to
always assume that camera is plugged in.

Signed-off-by: Kirill K. Smirnov <kirill.k.smirnov@gmail.com>