Specialized bandwidth setting for the musb TRIK controller. (#16)
TRIK musb cores don't support high bandwidth (HB) ISO transfers, therefore, we do not consider endpoints on a device that work with HB.

Some devices may request an incorrect or too large bandwidth, that TRIK does not support. Therefore, we use the largest bandwidth that TRIK supports for this device.

There is always a supported bandwidth for 320x240 resolution. For larger sizes, it may no longer be available, or the maximum possible one will be too slow.