Delete check cd gpio, fix i2c timeout (#19)
Removed the GPIO from the DT to detect the card in the card reader, since not all TRIK boards support this.

Fixed i2c timeout. Add bool parameter in DT to support recovery bus. This logic is used mostly for i2c0, i2c1 works stably.
Error is possible if there are no interruptions:
2.1 If the SCL or SDA are held in a low state, then they will be transferred to a high state and artificial clocking of the SCL is possible.
2.2 Also this can happen even with the bus running normally, there may be something wrong with the slave, because there are no interruptions coming from it. There is just a reinit (full reset) of the i2c module.

Add `reinit` file in sysfs for OV7670 so that it would be possible to connect the video module to the hot one and reinitialize it via i2c. The function also contains error handling logic when the i2c bus is stuck.